Make minimap its own file
